Add an allow-unsafe-code feature
This commit forbids unsafe code by default (#![deny(unsafe_code)]), but
adds an allow-unsafe-code feature that is enabled by default. This
mostly exists to satisfy anyone who detests `unsafe`. Its existence
proves that the unsafe code in this crate only exists for the FFI and
not for some more fundamental reason.

Fixes: https://github.com/psychon/x11rb/issues/59
Signed-off-by: Uli Schlachter <psychon@znc.in>